% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adrs}
\alias{admiral_adrs}
\title{Response Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3274 rows and 80 columns.
}
\source{
Derived from the \code{adsl}, \code{rs} and \code{tu} datasets using \code{{admiralonco}}
(\url{https://github.com/pharmaverse/admiralonco/blob/main/inst/templates/ad_adrs.R})
}
\usage{
admiral_adrs
}
\description{
An example response analysis dataset
}
\keyword{datasets}
