% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_confirmed_bor.R
\name{aval_resp}
\alias{aval_resp}
\title{Map Character Response Values to Numeric Values}
\usage{
aval_resp(arg)
}
\arguments{
\item{arg}{Character vector}
}
\value{
\code{1} if \code{arg} equals \code{"Y"}, \code{0} if \code{arg} equals \code{"N"}, \code{NA_real_} otherwise
}
\description{
Map character response values like \code{"PR"} or \code{"SD"} to numeric values.
}
\examples{

aval_resp(c("Y", "N", NA_character_))
}
\author{
Stefan Bundfuss
}
\keyword{user_utility}
