% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_aval_fun}
\alias{call_aval_fun}
\title{Creates \code{AVAL} from \code{AVALC} by Calling User Function}
\usage{
call_aval_fun(dataset, aval_fun)
}
\arguments{
\item{dataset}{Input dataset

The variable \code{AVALC} is expected.

\emph{Permitted Values}: a dataframe}

\item{aval_fun}{Function returning the \code{AVALC} values

The specified function must expect one argument expecting a character
vector and must return a numeric vector.

\emph{Permitted Values}: a function}
}
\value{
The input dataset with \code{AVAL} added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use
\code{admiraldev::process_set_values_to()} instead.

Create \code{AVAL} from \code{AVALC} by calling a function provided by the user. If
calling the function fails, the error is caught and a helpful error message
provided.
}
\details{
The new variable \code{AVAL} is set to \code{aval_fun(AVALC)}.
}
\author{
Stefan Bundfuss
}
\concept{deprecated}
\keyword{deprecated}
