% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pd.R
\name{date_source}
\alias{date_source}
\title{Create a \code{date_source} object}
\usage{
date_source(dataset_name, filter = NULL, date, set_values_to = NULL)
}
\arguments{
\item{dataset_name}{The name of the dataset, i.e. a string, used to search for
the date.}

\item{filter}{An unquoted condition for filtering \code{dataset}.}

\item{date}{A variable or an expression providing a date. A date or a
datetime can be specified. An unquoted symbol or expression is expected.}

\item{set_values_to}{Variables to be set}
}
\value{
An object of class \code{date_source}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The \code{date_source()} function has been
deprecated without replacement as all functions using the \code{source_pd}
argument are deprecated as well.

Create a \code{date_source} object as input for \code{source_pd} argument in
\code{{admiralonco}} functions.
}
\examples{

pd <- date_source(
  dataset_name = "adrs",
  date = ADT,
  filter = PARAMCD == "PD"
)
}
\seealso{
Other deprecated:
\code{\link[=derive_param_bor]{derive_param_bor()}},
\code{\link[=derive_param_clinbenefit]{derive_param_clinbenefit()}},
\code{\link[=derive_param_confirmed_bor]{derive_param_confirmed_bor()}},
\code{\link[=derive_param_confirmed_resp]{derive_param_confirmed_resp()}},
\code{\link[=derive_param_response]{derive_param_response()}},
\code{\link[=filter_pd]{filter_pd()}}
}
\concept{deprecated}
\keyword{deprecated}
