% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralonco-package.R
\docType{package}
\name{admiralonco-package}
\alias{admiralonco}
\alias{admiralonco-package}
\title{admiralonco: Oncology Extension Package for ADaM in 'R' Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Programming oncology specific Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in 'R'. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team (2021), \url{https://www.cdisc.org/standards/foundational/adam}). The package is an extension package of the 'admiral' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralonco/}
  \item \url{https://github.com/pharmaverse/admiralonco}
  \item Report bugs at \url{https://github.com/pharmaverse/admiralonco/issues}
}

}
\author{
\strong{Maintainer}: Stefan Bundfuss \email{stefan.bundfuss@external.roche.com}

Authors:
\itemize{
  \item Stefan Bundfuss \email{stefan.bundfuss@external.roche.com}
  \item Amit Jain
  \item Kiran Peddamudium
  \item Michael Hall
  \item Nathan Rees
  \item Tomoyuki Namai
  \item Ross Farrugia
}

Other contributors:
\itemize{
  \item Yirong Cao [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
  \item Bristol Myers Squibb [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
