% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralonco-package.R
\docType{package}
\name{admiralonco-package}
\alias{admiralonco}
\alias{admiralonco-package}
\title{admiralonco: Oncology extension package for ADaM in R Asset Library}
\description{
A toolbox for programming oncology specific Clinical Data Standards Interchange Consortium
    (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a
    mandatory part of any New Drug or Biologics License Application submitted to the
    United States Food and Drug Administration (FDA). Analysis derivations are
    implemented in accordance with the "Analysis Data Model Implementation Guide"
    (CDISC Analysis Data Model Team, 2021, <https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package>).
}
\author{
\strong{Maintainer}: Stefan Bundfuss \email{stefan.bundfuss@roche.com}

Authors:
\itemize{
  \item Ross Farrugia
  \item Amit Jain
  \item Liming (Catherine) Jin
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
  \item Amgen Inc. [copyright holder, funder]
}

}
\keyword{internal}
