% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_clinbenefit.R
\name{derive_param_clinbenefit}
\alias{derive_param_clinbenefit}
\title{Derive a Clinical Benefit Parameter}
\usage{
derive_param_clinbenefit(
  dataset,
  dataset_adsl,
  source_param,
  source_resp,
  source_pd,
  source_datasets,
  by_vars = NULL,
  reference_date = TRTSDT,
  ref_start_window = 28,
  set_values_to,
  subject_keys = vars(STUDYID, USUBJID)
)
}
\arguments{
\item{dataset}{Input dataset. This is the dataset to which the clinical
benefit rate parameter will be added.

The variables \code{PARAMCD}, \code{AVALC}, \code{ADT}, and those specified by the \code{by_vars}
parameter are expected.}

\item{dataset_adsl}{ADSL dataset used as input for populating \code{subject_keys}
in input dataset.

The variables specified by the \code{subject_keys} parameter and the \code{reference_date}
parameter are expected.}

\item{source_param}{Character name of parameter code in \code{source_dataset} that
represents overall disease response assessment for a subject at a given timepoint,
e.g. \code{"OVR"} or \code{"OVRLRESP"}.}

\item{source_resp}{A \code{date_source} object specifying the dataset, date variable,
and filter condition used to identify objective response status.}

\item{source_pd}{A \code{date_source} object specifying the dataset, date variable,
and filter condition used to identify disease progression.}

\item{source_datasets}{Named list of length 1 specifying the dataset from
which response information for deriving clinical benefit rate will be collected.}

\item{by_vars}{A named list returned by \code{vars()} used together with \code{subject_keys}
to identify groupings within which the earliest date of clinical benefit rate
status will be chosen.}

\item{reference_date}{Name of variable representing the index date for
\code{ref_start_window}. A variable providing a date. An unquoted symbol is expected.}

\item{ref_start_window}{Integer representing number of days from \code{reference_date}
that must elapse before an evaluable non-PD assessment counts toward determining
clinical benefit.}

\item{set_values_to}{A named list returned by \code{vars()} containing new variables
and their static value to be populated for the clinical benefit rate parameter
records, e.g. \code{vars(PARAMCD = "CBR", PARAM = "Clinical Benefit Rate")}.}

\item{subject_keys}{A named list returned by \code{vars()} containing variables
used to uniquely identify subjects.}
}
\value{
A data frame containing a parameter representing clinical benefit/
disease control.
}
\description{
Add a clinical benefit/disease control parameter to the input dataset.
}
\details{
Clinical benefit/disease control is first identified for looking for subjects
having objective response, and then derived for subjects that have at least
one evaluable non-PD response assessment after a specified amount of time
from a reference date.
}
\author{
Andrew Smith
}
\keyword{adrs}
\keyword{derivation}
