% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_confirmed_bor.R
\name{derive_param_confirmed_bor}
\alias{derive_param_confirmed_bor}
\title{Derive Confirmed Best Overall Response Parameter}
\usage{
derive_param_confirmed_bor(
  dataset,
  dataset_adsl,
  filter_source,
  source_pd = NULL,
  source_datasets,
  reference_date,
  ref_start_window,
  ref_confirm,
  max_nr_ne = 1,
  accept_sd = FALSE,
  set_values_to,
  subject_keys = vars(STUDYID, USUBJID)
)
}
\description{
Derive confirmed best overall response (BOR) parameter
}
\author{
Stefan Bundfuss
}
