% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_lasta.R
\name{derive_param_lasta}
\alias{derive_param_lasta}
\title{derive_param_lasta}
\usage{
derive_param_lasta(
  dataset,
  order = admiral::vars(USUBJID, ADT),
  by_vars = admiral::vars(USUBJID),
  filter_source = PARAMCD == "OVR" & ANL01FL == "Y",
  source_pd = NULL,
  source_datasets = list(adrs = adrs),
  subject_keys = admiral::vars(STUDYID, USUBJID),
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataframe to which the Last Disease Assessment
parameter will be added. The variables \code{PARAMCD}, \code{AVAL},
\code{AVALC}, \code{ADT}, and those specified in the \code{order} and
\code{by_vars} argument are expected.}

\item{order}{Sort order, after which the last record shall be taken by
the \code{by_vars} to determine Last Disease Assessment}

\item{by_vars}{Grouping variables, the last of which (ordered by \code{order})
shall be taken as the Last Disease Assessment record}

\item{filter_source}{Filter to be applied to the input dataframe to derive
the Last Disease Assessment}

\item{source_pd}{A \code{date_source} object specifying the dataset_name, date
variable, and filter condition used to identify disease
progression and subsequently used (if provided) to remove
all records after disease progression from \code{dataset}}

\item{source_datasets}{A named \code{list} containing the name of the dataframe
in which to search for the progressive disease as
defined in \code{source_pd}}

\item{subject_keys}{Variables to uniquely identify a subject, used by
\code{source_pd} to specify a date}

\item{set_values_to}{A named \code{list} returned by \code{vars()} containing new
variables and their static value to be populated for
the Last Disease Assessment records, e.g.
vars(PARAMCD = "LSTAC", PARAM = "Last Disease
Assessment Censored at First PD by Investigator")`}
}
\value{
The dataframe passed in the \code{dataset} argument with additonal columns
and/or rows as set in the \code{set_values_to} argument.
}
\description{
\preformatted{Add a Last Disease Assessment parameter for each unique `by_vars` to the 
input dataframe passed into the dataset argument.
}
}
\details{
Calculates the last disease assessment for subjects (censored if
required using argument source_pd which will remove all records in the
dataframe being passed that occur after the date in source_pd)
}
\examples{

\dontrun{
pd <- date_source(
  dataset_name = “adrs”,
  date         = ADT,
  filter       = PARAMCD == “PD” & AVALC == “Y”
)

derive_param_lasta(
  dataset,
  filter_source = PARAMCD == "OVR" & ANL01FL == "Y",
  source_pd = list(pd),
  source_datasets = list(adrs = adrs),
  set_values_to = vars(
    PARAMCD = "LSTAC",
    PARAM = "Last Disease Assessment Censored at First PD by Investigator",
    PARCAT1 = "Tumor Response",
    PARCAT2 = "Investigator",
    PARCAT3 = "Recist 1.1",
    ANL01FL = "Y")
)
}

}
\author{
Stephen Gormley
}
\keyword{ADRS}
