% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_lasta.R
\name{derive_param_lasta}
\alias{derive_param_lasta}
\title{derive_param_lasta}
\usage{
derive_param_lasta(
  dataset,
  dataset_adsl,
  filter_source,
  order = admiral::vars(STUDYID, USUBJID, ADT),
  by_vars = admiral::vars(STUDYID, USUBJID),
  source_pd = NULL,
  source_datasets = NULL,
  subject_keys = admiral::vars(STUDYID, USUBJID),
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataframe from which the Last Disease Assessment will be
be derived from and added to.

The columns \code{PARAMCD}, \code{ADT}, and \code{AVALC}and the columns specified in
\code{subject_keys} and \code{reference_date} are expected.

\emph{Permitted Values:} a \code{data.frame()} object

\emph{Required or Optional:} Required}

\item{dataset_adsl}{The ADSL input dataframe

The columns specified in \code{subject_keys} are expected.

\emph{Permitted Values:} a \code{data.frame()} object

\emph{Required or Optional:} Required}

\item{filter_source}{Filter to be applied to \code{dataset} to derive the
Last Disease Assessment

\emph{Required or Optional:} Required}

\item{order}{Sort order, after which the last record shall be taken by
the \code{by_vars} to determine Last Disease Assessment. Created
using \code{admiral::vars()}.

\emph{Permitted Values:} an \code{admiral::vars} object

\emph{Default:} \code{admiral::vars(STUDYID, USUBJID, ADT)}

\emph{Required or Optional:} Required}

\item{by_vars}{Grouping variables, the last of which (ordered by \code{order})
shall be taken as the Last Disease Assessment record. Created
using \code{admiral::vars()}.

\emph{Permitted Values:} an \code{admiral::vars} object

\emph{Default:} \code{admiral::vars(STUDYID, USUBJID)}

\emph{Required or Optional:} Required}

\item{source_pd}{Date of first progressive disease (PD)

If the parameter is specified, the observations of the input \code{dataset} for
deriving the new parameter are restricted to observations up to the
specified date. Observations at the specified date are included. For
subjects without first PD date all observations are take into account.

\emph{Permitted Values:} a \code{date_source} object (see \code{admiral::date_source()}
for details)

\emph{Default:} \code{NULL}

\emph{Required or Optional:} Optional}

\item{source_datasets}{Source dataframe to be used to calculate the
first PD date

A named list of dataframes is expected (although for BoR) only one dataframe is
needed. It links the \code{dataset_name} from \code{source_pd} with an existing dataframe.

For example if \code{source_pd = pd_date} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{pd_date <- date_source(
  dataset_name = "adrs",
  date = ADT,
  filter = PARAMCD == PD
)
}\if{html}{\out{</div>}}

and the actual response dataframe in the script is \code{myadrs}, \code{source_datasets = list(adrs = myadrs)} should be specified.

\emph{Required or Optional:} Optional}

\item{subject_keys}{Variables to uniquely identify a subject

A list of symbols created using \code{admiral::vars()}.

\emph{Default:} \code{admiral::vars(STUDYID, USUBJID)}

\emph{Required or Optional:} Required}

\item{set_values_to}{Variables to set

A named list returned by \code{vars()} defining the variables to be set for the
new parameter, e.g. \code{vars(PARAMCD = "LSTAC", PARAM = "Last Disease Assessment Censored at First PD by Investigator")} is expected. The values
must be symbols, character strings, numeric values, or \code{NA}.

\emph{Required or Optional:} Required}
}
\value{
The dataframe passed in the \code{dataset} argument with additional
columns and/or rows as set in the \code{set_values_to} argument.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Add a Last Disease Assessment parameter for each unique `by_vars` to the
input dataframe passed into the dataset argument.
}\if{html}{\out{</div>}}
}
\details{
Calculates the last disease assessment.
Records after PD can be removed using the source_pd and source_datasets
arguments.
}
\examples{

\dontrun{

library(magrittr)

adsl <- tibble::tribble(
  ~USUBJID, ~TRTSDT,                      ~EOSDT,
  "01",     lubridate::ymd("2020-12-06"), lubridate::ymd("2022-03-06"),
  "02",     lubridate::ymd("2021-01-16"), lubridate::ymd("2022-02-03"),
  "03",     lubridate::ymd("2021-01-09"), lubridate::ymd("2021-02-24"),
  "04",     lubridate::ymd("2021-04-21"), lubridate::ymd("2021-09-15")
) \%>\%
  dplyr::mutate(STUDYID = "a_study_id")

adrs <- tibble::tribble(
  ~USUBJID, ~PARAMCD, ~AVAL, ~AVALC, ~ASEQ, ~ADT,
  "01", "RSP", NA, "Y", 1, lubridate::ymd("2021-04-08"),
  "02", "RSP", NA, "N", 1, lubridate::ymd("2021-05-07"),
  "03", "RSP", NA, "N", 1, NA,
  "04", "RSP", NA, "N", 1, NA,
  "01", "PD",  NA, "N", 1, NA,
  "02", "PD",  NA, "Y", 1, lubridate::ymd("2021-05-07"),
  "03", "PD",  NA, "N", 1, NA,
  "04", "PD",  NA, "N", 1, NA,
  "01", "OVR", 3, "SD", 1, lubridate::ymd("2021-03-07"),
  "01", "OVR", 2, "PR", 1, lubridate::ymd("2021-04-08"),
  "02", "OVR", 3, "SD", 1, lubridate::ymd("2021-03-07"),
  "02", "OVR", NA, NA, 1, lubridate::ymd("2021-04-07"),
  "02", "OVR", 6, "PD", 1, lubridate::ymd("2021-05-07"),
  "03", "OVR", 3, "SD", 1, lubridate::ymd("2021-01-30"),
  "03", "OVR", 3, "SD", 2, lubridate::ymd("2021-01-30"),
  "04", "OVR", NA, "NE", 1, lubridate::ymd("2021-05-21"),
  "04", "OVR", 5, "NON-PD", 1, lubridate::ymd("2021-06-30"),
  "04", "OVR", NA, "NE", 1, lubridate::ymd("2021-07-24"),
  "04", "OVR", NA, "ND", 1, lubridate::ymd("2021-09-30"),
) \%>\%
  dplyr::mutate(STUDYID = "a_study_id")

pd <- date_source(
  dataset_name = adrs,
  date         = ADT,
  filter       = PARAMCD == 'PD' & AVALC == 'Y'
)

derive_param_lasta(
  dataset,
  filter_source = PARAMCD == "OVR" & ANL01FL == "Y",
  source_pd = list(pd),
  source_datasets = list(adrs = adrs),
  set_values_to = vars(
    PARAMCD = "LSTAC",
    PARAM = "Last Disease Assessment Censored at First PD by Investigator",
    PARCAT1 = "Tumor Response",
    PARCAT2 = "Investigator",
    PARCAT3 = "Recist 1.1",
    ANL01FL = "Y")
)
}

}
\author{
Stephen Gormley
}
\keyword{ADRS}
