% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_lasta.R
\name{derive_param_lasta}
\alias{derive_param_lasta}
\title{Adds a Parameter for the Last Disease Assessment}
\usage{
derive_param_lasta(
  dataset,
  filter_source,
  order = vars(STUDYID, USUBJID, ADT),
  by_vars = vars(STUDYID, USUBJID),
  source_pd = NULL,
  source_datasets = NULL,
  subject_keys = vars(STUDYID, USUBJID),
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataframe from which the Last Disease Assessment will be
be derived from and added to.

The columns \code{PARAMCD}, \code{ADT}, and \code{AVALC}and the columns specified in
\code{subject_keys} and \code{reference_date} are expected.

\emph{Permitted Values:} a \code{data.frame()} object

\emph{Required or Optional:} Required}

\item{filter_source}{Filter to be applied to \code{dataset} to derive the
Last Disease Assessment

\emph{Required or Optional:} Required}

\item{order}{Sort order, after which the last record shall be taken by
the \code{by_vars} to determine Last Disease Assessment. Created
using \code{vars()}.

\emph{Permitted Values:} an \code{vars} object

\emph{Default:} \code{vars(STUDYID, USUBJID, ADT)}

\emph{Required or Optional:} Required}

\item{by_vars}{Grouping columns, the last of which (ordered by \code{order})
shall be taken as the Last Disease Assessment record. Created
using \code{vars()}.

\emph{Permitted Values:} an \code{vars} object

\emph{Default:} \code{vars(STUDYID, USUBJID)}

\emph{Required or Optional:} Required}

\item{source_pd}{Date of first progressive disease (PD)

If the parameter is specified, the observations of the input \code{dataset} for
deriving the new parameter are restricted to observations up to the
specified date. Observations at the specified date are included. For
subjects without first PD date all observations are take into account.

\emph{Permitted Values:} a \code{date_source} object (see \code{date_source()}
for details)

\emph{Default:} \code{NULL}

\emph{Required or Optional:} Optional}

\item{source_datasets}{Source dataframe to be used to calculate the
first PD date

A named list of dataframes is expected (although for BoR) only one dataframe is
needed. It links the \code{dataset_name} from \code{source_pd} with an existing dataframe.

For example if \code{source_pd = pd_date} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{pd_date <- date_source(
  dataset_name = "adrs",
  date = ADT,
  filter = PARAMCD == PD
)
}\if{html}{\out{</div>}}

and the actual response dataframe in the script is \code{myadrs}, \code{source_datasets = list(adrs = myadrs)} should be specified.

\emph{Required or Optional:} Optional}

\item{subject_keys}{Columns to uniquely identify a subject

A list of symbols created using \code{vars()}.

\emph{Default:} \code{vars(STUDYID, USUBJID)}

\emph{Required or Optional:} Required}

\item{set_values_to}{Columns to set

A named list returned by \code{vars()} defining the columns to be set for the
new parameter, e.g. \code{vars(PARAMCD = "LSTAC", PARAM = "Last Disease Assessment Censored at First PD by Investigator")} is expected. The values
must be symbols, character strings, numeric values, or \code{NA}.

\emph{Required or Optional:} Required}
}
\value{
The dataframe passed in the \code{dataset} argument with additional
columns and/or rows as set in the \code{set_values_to} argument.ß
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Derives Last Disease Assessment (optionally) up to first Progressive Disease
}\if{html}{\out{</div>}}
}
\details{
Calculates the last disease assessment by accessing the last record
defined in \code{by_vars} after it has been arranged using the \code{order} argument.

Creates the new parameter record with AVAL/AVALC taken from the source
record and ADT as the last ADT.

Records after PD can be removed using the source_pd and source_datasets
arguments.

Note: All columns from the input dataset are kept. Subjects with no records
in the input dataset (after the filter is applied) all records are kept from
ADSL which are also in the input dataset.  Columns which are not be populated
for the new parameter/populated differently (e.g., RSSTRESC, VISIT, ANLzzFL)
should be overwritten using the \code{set_values_to} parameter.
}
\examples{

library(dplyr)
library(lubridate)
library(admiral)
library(tibble)

adsl <- tribble(
  ~USUBJID, ~TRTSDT,                      ~EOSDT,
  "01",     lubridate::ymd("2020-12-06"), lubridate::ymd("2022-03-06"),
  "02",     lubridate::ymd("2021-01-16"), lubridate::ymd("2022-02-03"),
  "03",     lubridate::ymd("2021-01-09"), lubridate::ymd("2021-02-24"),
  "04",     lubridate::ymd("2021-04-21"), lubridate::ymd("2021-09-15")
) \%>\%
  mutate(STUDYID = "a_study_id")

adrs <- tribble(
  ~USUBJID, ~PARAMCD, ~AVAL, ~AVALC, ~ASEQ, ~ADT,
  "01", "RSP", NA, "Y", 1, lubridate::ymd("2021-04-08"),
  "02", "RSP", NA, "N", 1, lubridate::ymd("2021-05-07"),
  "03", "RSP", NA, "N", 1, NA,
  "04", "RSP", NA, "N", 1, NA,
  "01", "PD", NA, "N", 1, NA,
  "02", "PD", NA, "Y", 1, lubridate::ymd("2021-05-07"),
  "03", "PD", NA, "N", 1, NA,
  "04", "PD", NA, "N", 1, NA,
  "01", "OVR", 3, "SD", 1, lubridate::ymd("2021-03-07"),
  "01", "OVR", 2, "PR", 1, lubridate::ymd("2021-04-08"),
  "02", "OVR", 3, "SD", 1, lubridate::ymd("2021-03-07"),
  "02", "OVR", NA, NA, 1, lubridate::ymd("2021-04-07"),
  "02", "OVR", 6, "PD", 1, lubridate::ymd("2021-05-07"),
  "03", "OVR", 3, "SD", 1, lubridate::ymd("2021-01-30"),
  "03", "OVR", 3, "SD", 2, lubridate::ymd("2021-01-30"),
  "04", "OVR", NA, "NE", 1, lubridate::ymd("2021-05-21"),
  "04", "OVR", 5, "NON-PD", 1, lubridate::ymd("2021-06-30"),
  "04", "OVR", NA, "NE", 1, lubridate::ymd("2021-07-24"),
  "04", "OVR", NA, "ND", 1, lubridate::ymd("2021-09-30"),
) \%>\%
  mutate(STUDYID = "a_study_id")

pd <- date_source(
  dataset_name = adrs,
  date         = ADT,
  filter       = PARAMCD == "PD" & AVALC == "Y"
)

derive_param_lasta(
  dataset,
  filter_source = PARAMCD == "OVR" & ANL01FL == "Y",
  source_pd = pd,
  source_datasets = list(adrs = adrs),
  set_values_to = vars(
    PARAMCD = "LSTAC",
    PARAM = "Last Disease Assessment Censored at First PD by Investigator",
    PARCAT1 = "Tumor Response",
    PARCAT2 = "Investigator",
    PARCAT3 = "Recist 1.1",
    ANL01FL = "Y"
  )
)
}
\author{
Stephen Gormley
}
\keyword{ADRS}
