% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_lasta.R
\name{derive_param_lasta}
\alias{derive_param_lasta}
\title{Adds a Parameter for the Last Disease Assessment}
\usage{
derive_param_lasta(
  dataset,
  filter_source,
  order = vars(ADT),
  source_pd = NULL,
  source_datasets = NULL,
  subject_keys = vars(STUDYID, USUBJID),
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataframe from which the Last Disease Assessment will be
be derived from and added to.

The column \code{PARAMCD} and the columns specified in \code{subject_keys} and
\code{order} are expected.

\emph{Permitted Values:} a \code{data.frame()} object

\emph{Required or Optional:} Required}

\item{filter_source}{Filter to be applied to \code{dataset} to derive the
Last Disease Assessment

\emph{Required or Optional:} Required}

\item{order}{Sort order, after which the last record shall be taken by
the \code{subject_keys} to determine Last Disease Assessment. Created
using \code{vars()}.

\emph{Permitted Values:} list of variables or \verb{desc(<variable>)} function calls
created by \code{vars()}, e.g., \code{vars(ADT, desc(AVAL))}

\emph{Default:} \code{vars(ADT)}

\emph{Required or Optional:} Required}

\item{source_pd}{Date of first progressive disease (PD)

If the parameter is specified, the observations of the input \code{dataset} for
deriving the new parameter are restricted to observations up to the
specified date. Observations at the specified date are included. For
subjects without first PD date all observations are take into account.

\emph{Permitted Values:} a \code{date_source} object (see \code{date_source()}
for details)

\emph{Default:} \code{NULL}

\emph{Required or Optional:} Optional}

\item{source_datasets}{Source dataframe to be used to calculate the
first PD date

A named list of dataframes is expected (although for BOR) only one dataframe is
needed. It links the \code{dataset_name} from \code{source_pd} with an existing dataframe.

For example if \code{source_pd = pd_date} with\if{html}{\out{<div class="r">}}\preformatted{pd_date <- date_source(
  dataset_name = "adrs",
  date = ADT,
  filter = PARAMCD == PD
)
}\if{html}{\out{</div>}}

and the actual response dataframe in the script is \code{myadrs}, \code{source_datasets = list(adrs = myadrs)} should be specified.

\emph{Required or Optional:} Optional}

\item{subject_keys}{Columns to uniquely identify a subject

A list of symbols created using \code{vars()}.

\emph{Default:} \code{vars(STUDYID, USUBJID)}

\emph{Required or Optional:} Required}

\item{set_values_to}{Columns to set

A named list returned by \code{vars()} defining the columns to be set for the
new parameter, e.g. \code{vars(PARAMCD = "LSTAC", PARAM = "Last Disease Assessment Censored at First PD by Investigator")} is expected. The values
must be symbols, character strings, numeric values, or \code{NA}.

\emph{Required or Optional:} Required}
}
\value{
The dataframe passed in the \code{dataset} argument with additional
columns and/or rows as set in the \code{set_values_to} argument.
}
\description{
Adds a parameter for the last disease assessment (optionally) up to first progressive
disease
}
\details{
Calculates the last disease assessment by accessing the last record
defined in \code{subject_keys} after it has been arranged using the \code{order} argument.

Creates a new parameter record (keeping all columns passed) from the last source
record (i.e. the last record defined in \code{subject_keys} after it has been arranged
using the \code{order} argument). One new record for each subject in the filtered
input \code{dataset} is added to the input \code{dataset}.

Records after PD can be removed using the \code{source_pd} and \code{source_datasets}
arguments.
}
\examples{

library(dplyr)
library(lubridate)
library(admiral)
library(tibble)
library(magrittr)

adsl <- tribble(
  ~USUBJID, ~TRTSDT,                      ~EOSDT,
  "01",     ymd("2020-12-06"), ymd("2022-03-06"),
  "02",     ymd("2021-01-16"), ymd("2022-02-03"),
  "03",     ymd("2021-01-09"), ymd("2021-02-24"),
  "04",     ymd("2021-04-21"), ymd("2021-09-15")
) \%>\%
  mutate(STUDYID = "a_study_id")

adrs <- tribble(
  ~USUBJID, ~PARAMCD, ~AVAL, ~AVALC, ~ASEQ, ~ADT, ~ANL01FL,
  "01", "RSP", NA, "Y", 1, ymd("2021-04-08"), NA,
  "02", "RSP", NA, "N", 1, ymd("2021-05-07"), NA,
  "03", "RSP", NA, "N", 1, NA, NA,
  "04", "RSP", NA, "N", 1, NA, NA,
  "01", "PD", NA, "N", 1, NA, NA,
  "02", "PD", NA, "Y", 1, ymd("2021-05-07"), NA,
  "03", "PD", NA, "N", 1, NA, NA,
  "04", "PD", NA, "N", 1, NA, NA,
  "01", "OVR", 3, "SD", 1, ymd("2021-03-07"), "Y",
  "01", "OVR", 2, "PR", 1, ymd("2021-04-08"), "Y",
  "02", "OVR", 3, "SD", 1, ymd("2021-03-07"), "Y",
  "02", "OVR", NA, NA, 1, ymd("2021-04-07"), NA,
  "02", "OVR", 6, "PD", 1, ymd("2021-05-07"), "Y",
  "03", "OVR", 3, "SD", 1, ymd("2021-01-30"), NA,
  "03", "OVR", 3, "SD", 2, ymd("2021-01-30"), "Y",
  "04", "OVR", NA, "NE", 1, ymd("2021-05-21"), "Y",
  "04", "OVR", 5, "NON-PD", 1, ymd("2021-06-30"), "Y",
  "04", "OVR", NA, "NE", 1, ymd("2021-07-24"), "Y",
  "04", "OVR", NA, "ND", 1, ymd("2021-09-30"), "Y"
) \%>\%
  mutate(STUDYID = "a_study_id")

pd <- date_source(
  dataset_name = "adrs",
  date         = ADT,
  filter       = PARAMCD == "PD" & AVALC == "Y"
)

derive_param_lasta(
  adrs,
  filter_source = PARAMCD == "OVR" & ANL01FL == "Y",
  source_pd = pd,
  source_datasets = list(adrs = adrs),
  set_values_to = vars(
    PARAMCD = "LSTAC",
    PARAM = "Last Disease Assessment Censored at First PD by Investigator",
    PARCAT1 = "Tumor Response",
    PARCAT2 = "Investigator",
    PARCAT3 = "Recist 1.1",
    ANL01FL = "Y"
  )
) \%>\%
  filter(PARAMCD == "LSTAC")
}
\author{
Stephen Gormley
}
\keyword{adrs}
